Refactor message send logic
