+++ src/Juridical.Message.Worker/Program.cs
@@ -1,7 +1,8 @@
 using Juridical.Message.Worker;
+using Juridical.Message.Worker.Workers;
 
 IHost host = Host.CreateDefaultBuilder(args)
-    .ConfigureServices(services => { services.AddHostedService<Worker>(); })
+    .ConfigureServices(services => { services.AddHostedService<MessageWorker>(); })
     .Build();
 
 await host.RunAsync();
